changed all ids to BIGINT in sql scripts

+++ backend/sql/deploy.sql
@@ -6,7 +6,7 @@ USE `oilwhere`;
 -- Creation of Customer table to hold details about customer, including their name and email address
 DROP TABLE IF EXISTS `customer`;
 CREATE TABLE IF NOT EXISTS `customer` (
-  `customer_id` INT NOT NULL AUTO_INCREMENT,
+  `customer_id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` varchar(32) NOT NULL,
   `email` varchar(255),
   PRIMARY KEY (`customer_id`)
@@ -15,11 +15,11 @@ CREATE TABLE IF NOT EXISTS `customer` (
 -- Create a temporary table with sale_date as VARCHAR to help convert date to appropriate format and dropped afterwards
 DROP TABLE IF EXISTS `purchase_history_temp`;
 CREATE TABLE IF NOT EXISTS `purchase_history_temp` (
-  `purchase_id` INT NOT NULL AUTO_INCREMENT,
+  `purchase_id` BIGINT NOT NULL AUTO_INCREMENT,
   `sale_date` VARCHAR(10) NOT NULL,  -- Store date as VARCHAR first
   `sale_type` VARCHAR(255),
   `digital`  VARCHAR(255),
-  `customer_id` INT NOT NULL,
+  `customer_id` BIGINT NOT NULL,
   `zipcode` INT(6),
   `shipping_method` VARCHAR(255),
   `product` VARCHAR(255),
@@ -58,11 +58,11 @@ FROM `purchase_history_temp`;
 -- Creation of purchase_history table
 DROP TABLE IF EXISTS `purchase_history`;
 CREATE TABLE IF NOT EXISTS `purchase_history` (
-  `purchase_id` INT NOT NULL AUTO_INCREMENT,
+  `purchase_id` BIGINT NOT NULL AUTO_INCREMENT,
   `sale_date` date NOT NULL,
   `sale_type` varchar(255),
   `digital`  varchar(255),
-  `customer_id` int NOT NULL,
+  `customer_id` BIGINT NOT NULL,
   `zipcode` int(6),
   `shipping_method` varchar(255),
   `product` varchar(255),
@@ -85,7 +85,7 @@ CREATE TABLE IF NOT EXISTS `customer_spending_ranked` (
 -- Creation of user table to hold account personal details
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE IF NOT EXISTS `user` (
-  `user_id` INT NOT NULL AUTO_INCREMENT,
+  `user_id` BIGINT NOT NULL AUTO_INCREMENT,
   `username` varchar(255) NOT NULL,
   `role` varchar(32) NOT NULL,
   `password` varchar(32) NOT NULL,
@@ -95,7 +95,7 @@ CREATE TABLE IF NOT EXISTS `user` (
 -- Creation of NewsLetter table to hold all newsletter details
 DROP TABLE IF EXISTS `newsletter`;
 CREATE TABLE IF NOT EXISTS `newsletter` (
-  `design_id` INT NOT NULL AUTO_INCREMENT,
+  `design_id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` varchar(32) NOT NULL,
   `content` text NOT NULL,
   PRIMARY KEY (`design_id`)
